+++ bin/index.js
@@ -1,12 +1,9 @@
 #!/usr/bin/env node
 import { Command } from "commander";
 import log from "../src/utils/log/log.js";
-import { createRequire } from "module";
-const require = createRequire(import.meta.url);
-const configDefault = require("../src/config.default.json");
 
 // show header
-log("infoStart", "header", configDefault);
+log("infoStart", "header");
 
 // commander program
 const program = new Command();
@@ -15,7 +12,7 @@ program
   .description(
     "Website-Text-Translator: Automatically translate the text in HTML files to multiple languages, store translations in JSON files and detect changes."
   )
-  .version("0.1.0")
+  .version("1.0.0")
   .command("start", "Start the website-text-translator program.", {
     isDefault: true,
   })