+++ src/utils/getHtmlData.js
@@ -1,13 +1,13 @@
 import fs from "fs";
 import log from "./log/log.js";
 
-export default (fileName, config) => {
+export default (fileName) => {
   try {
     const rawData = fs.readFileSync(`${fileName}`, "utf8");
-    log("htmlRead", "info", config, [fileName]);
+    log("htmlRead", "info", [fileName]);
     return rawData;
   } catch (e) {
-    log("htmlReadFail", "error", config, [e]);
+    log("htmlReadFail", "error", [e]);
     return null;
   }
 };