+++ src/utils/log/messages.js
@@ -1,23 +1,28 @@
 export default {
-  infoStart:
-    "Website-Text-Translator, by liondigits\x1b[0m\nhttps://liondigits.com",
-  infoEnd: "All done!",
+  title: "Website-Text-Translator",
+  author: "Author: Joris Raymaekers",
+  website: "Visit: https://liondigits.com",
+  github: "Github repo: https://github.com/jorishr/website-text-translator",
+  programEnd: "All done!",
   backupStart: "Safe mode: ON. Starting backup...",
   backupFail: "Backup failed. No HTML and JSON file(s) found in source folder.",
+  backupDone: "Backup completed.",
   htmlStart: "Start processing HTML file: ${0}",
-  htmlRead: "Read HTML file: ${0}",
-  htmlList: "Found HTML file(s) to process:\n${0}\n",
+  htmlRead: "Parsed HTML file: ${0}",
+  htmlList: "Found HTML file(s) to process:\n${0}",
   htmlReadFail: "Unable to load HTML data from folder:\n\n\t${0}",
   htmlNotFound:
     "Program terminated. No HTML file(s) found in folder\n\nHint:\n\tMake sure you have at least one HTML file in the source folder: ${0}\n\tYou can change the default folder path by running the config command.",
   htmlDone: "Done processing HTML file: ${0}",
-  htmlEnd: "Done processing all HTML and JSON base file(s)!",
+  htmlEnd: "Done processing all HTML and JSON base language file(s)!",
   elementsFound:
-    "Found a total of ${0} HTML elements with an existing data-id in this HTML file.\n\n${1} new HTML elements were found.\n",
+    "Found a total of ${0} HTML elements with an existing data-id.\n\n${1} new HTML elements were found.\n",
   newElemsStart: "Start processing new HTML elements...",
   newElemsDone: "Done processing new HTML elements!",
-  obsoleteKeys:
-    "The following keys are obsolete and will be removed from the base JSON file: ${0}",
+  obsoleteKeysStart: "Checking for obsolete keys to be removed...",
+  obsoleteKeysFound:
+    "The following keys are obsolete and will be removed from the base language JSON file and target language file(s): ${0}",
+  obsoleteKeysNoneFound: "No obsolete keys were detected.",
   writeFileStart: "Start writing to file ${0}...",
   mkdir: "Directory ${0} does not exist. Creating...",
   getFilesFail: "Unexpected error while reading folder content:\n\n\t${0}",
@@ -26,32 +31,54 @@ export default {
   backupWriteFail: "Unexpected error while writing to backup file:\n\n\t${0}",
   missingLang:
     "No base or target languages set. Run the config command or add a config file. See documentation for more info.",
+  baseLangFileError:
+    "The base language file could not be read or is missing. Check the destination folder. Did you delete the file? Is the file corrupted? Try restoring the base language JSON file from your backup folder. If this was intentional, strip the text-id's from the html files using the wtt-strip command and start clean.",
   translateStart: "Start processing translation file(s)...",
+  translateEnd: "Done processing translations.",
   translateDisabled: "Processing translation files is disabled.\n",
   translateNoKeys:
-    "No changes found in base files. No translation changes needed.",
+    "No changes found in base files. Skipping translation tasks.\n",
+  translateError1: "Translation error while processing ${0} for language ${1}.",
+  translateError2: "Google Translate API error.\n\n\t${0}",
   langFileExists:
     "Found existing translation file for language: ${0}. Applying updates...",
   langFileNew:
     "No existing translation data found for language: ${0}. Creating new file...",
+  langFileDone: "Done processing translation file for language: ${0}",
+  addLangTargetsStart: "Start adding translation JSON file for language: ${0}.",
+  addLangTargetsDone: "Done adding new target language(s): ${0}.",
   fetchTranslation:
     "Fetching translations for ${0} text strings from the Google Translate API for language: ${1}...`",
-  txtUpdateException:
+  textUpdateException:
     "WARNING: Unsupported target value while setting keys with text updates. This should not happen. Please report his issue.",
-  getTxtChanges: "\nSearching for text changes in ${0}...",
-  getTxtChangesException:
+  getTextChanges: "Searching for text changes in ${0}...",
+  getTextChangesException:
     "WARNING: Unsupported target value while getting keys with text updates. This should not happen. Please report his issue.",
-  noChangesFound: "No changes found in ${0}.",
-  txtChange: "txt value for txt-id ${0} has changed",
-  attrChange: "txt value for ${0} attribute txt-id ${1} has changed",
-  txtAdded: "Added txt-id ${0} to ${1} element",
+  noChangesFound: "No changes found in ${0}.\n",
+  textChange: "text value for text-id ${0} has changed",
+  attrChange: "text value for ${0} attribute text-id ${1} has changed",
+  textAdded: "Added text-id ${0} to ${1} element",
   attrAdded: "Added ${0}=${1} to ${2} element",
   dryRun: "Translations are disabled in dry run mode.",
   dryRunStart:
     "Dry run mode: ON.\nNo files will be written.\nNo translations will be fetched from the Google Translate API.",
-  jsonRead: "The base JSON file contains ${0} strings.",
-  jsonNotFound: "No base file JSON data found. Continue without...",
-  startStrip: "\nStripping txt-id's from HTML file: ${0}",
+  baseLangDataRead:
+    "\nThe base language JSON file contains ${0} keys with a text value.\n",
+  baseLangDataNotFound:
+    "No base language JSON file found (${0}). A new one will be generated...\n",
+  jsonLangFilesFound: "The following language file(s) will be deleted: ${0}",
+  jsonLangFilesNotFound:
+    "No JSON files found to delete in folder: ${0}. Did you delete them manually?",
+  jsonLangFileDel: "${0} Deleted.",
+  jsonLangFileDelError: "Error deleting file: ${0}.\n\n${1}",
+  customKeysStart: "Checking custom key(s): ${0}.",
+  customKeysNotFound:
+    "[WARNING] Custom key not found in base language JSON file: Missing key: ${0}.\nThis key will be ignored. Check documentation for more details.",
+  customKeysAdded: "Updated JSON language file ${0} with key(s): ${1}.",
+  customKeysNoChanges: "No new custom keys detected.\n",
+  customKeysDone: "Done processing custom keys.",
+  stripStart: "\nStripping text-id's from HTML file: ${0}",
+  stripEnd: "Done stripping text-id's from HTML file: ${0}",
   configReadError: "Error while reading config file:\n ${0}",
   config: {
     welcome: "Welcome to the Website-Text-Translator configuration wizard.",
@@ -61,10 +88,10 @@ export default {
         "1. Please enter the base language of your project.\n\nThis is the language that you used to write the text in your HTML files. Use two-letter language codes, e.g. 'en' for English, 'de' for German, 'fr' for French, etc.",
       prompt1: "\x1b[3mBase language:\x1b[0m ",
       info2:
-        "2. Please enter the target languages to translate to.\n\nYou can specify multiple languages by separating them with a comma. Use two-letter language codes, e.g. 'en, de, fr' for English, German and French.",
+        "2. Please enter the target languages to translate to.\n\nYou can specify multiple languages by separating them with a comma. Use two-letter language codes, e.g. en, de, fr for English, German and French.",
       prompt2: "\x1b[3mTarget language(s):\x1b[0m ",
       inValid:
-        "Invalid input. Please try again. Use two-letter language codes: e.g. 'en' for English, 'de' for German, 'fr' for French, etc. Target languages can be separated by a comma, e.g. 'en, de, fr' for English, German and French.",
+        "Invalid input. Please try again. Use two-letter language codes: e.g. 'en' for English, 'de' for German, 'fr' for French, etc. Target languages can be separated by a comma, e.g. en, de, fr for English, German and French.",
       inValidEqual: "Base and target languages can't be the same.",
     },
     folders: {